Perform yarn upgrade
The commit performs a `yarn upgrade` of the framework to better
represent what downstream applications pull with our version ranges, and
to resolve known security vulnerabilities which were pulled by our
lockfile. The changes also make sure that our declared ranges for
dependencies are correct and fixes any compilation errors.

nan in version 2.19.0 leads to issues during electron rebuild. This is
why we will fix nan to 2.18.0 for now, and cpu-features to 0.0.9, since
this would like to have nan2.19.0

Also removes node 16 from ci builds, since this node version is not
supported any more.

Contributed on behalf of STMicroelectronics